Restart serve.py
