+++ crosscompute/constants.py
@@ -4,3 +4,5 @@ from os.path import expanduser
 
 FOLDER = getenv(
     'CROSSCOMPUTE_FOLDER', expanduser('~/.crosscompute'))
+HOST = '127.0.0.1'
+PORT = 7000
